LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/position-anchor-003-expected.txt


PASS Layout is updated on `position-anchor` changes
